refactor: api 조건분기 추가
