+++ src/api/getMyPage/GetMyPage.ts
@@ -1,4 +1,4 @@
-import api from '../apiClient';
+import api from '../index';
 
 export const GetMyPage = async () => {
   try {
