+++ src/api/interest/interestAPI.ts
@@ -1,4 +1,4 @@
-import api from '../apiClient';
+import api from '../index';
 
 export const registerInterest = async (interestTypes: string[]) => {
   try {
