+++ src/api/interest/updateInterestAPI.ts
@@ -1,4 +1,4 @@
-import api from '../apiClient';
+import api from '../index';
 
 export const updateInterest = async (interestTypes: string[]) => {
   try {
