+++ src/api/logout/logoutAPI.ts
@@ -1,4 +1,4 @@
-import api from '../apiClient';
+import api from '../index';
 
 export const logoutAPI = async () => {
   return await api.post('/auth/logout');
